event_module/requirements.txt
amqp==5.1.1
asgiref==3.5.0
billiard==3.6.4.0
celery==5.2.6
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Django==3.2.13
django-celery-beat==2.2.1
django-celery-results==2.3.1
django-environ==0.8.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
kombu==5.2.4
prompt-toolkit==3.0.29
psycopg2==2.9.3
PyJWT==2.3.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-http-client==3.3.7
pytz==2022.1
sendgrid==6.9.7
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
tzdata==2022.1
vine==5.0.0
wcwidth==0.2.5
